\name{check_subject_pattern}
\alias{check_subject_pattern}
\title{check subject pattern}
\description{Error if subject or \code{pattern} incorrect type.}
\usage{check_subject_pattern(subject.vec, pattern)}
\arguments{
  \item{subject.vec}{
}
  \item{pattern}{
}
}



\author{Toby Dylan Hocking}





