\name{names_or_error}
\alias{names_or_error}
\title{names or error}
\description{Extract capture group names. Stop with an error if there are no
capture groups, or if there are any capture groups without names.}
\usage{names_or_error(vec.with.attrs, 
    pattern)}
\arguments{
  \item{vec.with.attrs}{Output from g?regexpr.}
  \item{pattern}{
}
}

\value{Character vector.}

\author{Toby Dylan Hocking}





