\name{only_captures}
\alias{only_captures}
\title{only captures}
\description{Extract capture group columns from \code{match.mat}, stop if any are
un-named, and assign optional groups to "".}
\usage{only_captures(match.mat, 
    pattern)}
\arguments{
  \item{match.mat}{match.mat }
  \item{pattern}{pattern }
}



\author{Toby Dylan Hocking}





